python/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["John Sungjin Kim <jsjkim00@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
anthropic = "^0.3.6"
ipykernel = "^6.25.0"
ipywidgets = "^8.0.7"
langchain = "^0.0.247"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
